Задача 2/4, исправление ошибок после переноса вопросов в QuestionFactory
