+++ MovieQuiz/Models/QuizStepViewModel.swift
@@ -1,0 +1,15 @@
+//
+//  QuizStepViewModel.swift
+//  MovieQuiz
+//
+//  Created by Dinara on 11.07.2023.
+//
+
+import Foundation
+import UIKit
+
+struct QuizStepViewModel {
+    let image: UIImage
+    let question: String
+    let questionNumber: String
+}
